Implement po4a translation framework for the LinuxCNC documentation
Handle both adoc and manual pages.  All texts are grouped into one POT
file, as splitting them into groups did not work with the version of po4a
I got.  This feature is implemented in the latest po4a edition and we
should consider splitting into several POT files when this is available.